python/viralkids/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.0.0
agno>=0.1.0
python-multipart>=0.0.6

